Clean up duplicate .well-known directories
- Remove incorrect /app/api/.well-known/ directory
- Keep only /app/.well-known/ for proper RFC 5785 compliance
- Copy latest oauth-protected-resource updates to correct location
- Ensure .well-known endpoints are at domain root
