+++ app/.well-known/oauth-protected-resource/route.ts
@@ -1,63 +1,93 @@
 /**
- * OAuth 2.0 Protected Resource Metadata for MCP Server
- * RFC 8707 - Resource Indicators for OAuth 2.0
- * Required by MCP Authorization Specification
+ * OAuth 2.0 Protected Resource Metadata Endpoint
+ * RFC 8705 - OAuth 2.0 Resource Server Metadata
+ * Tells clients where the protected MCP resources are located
  */
 
-import { NextRequest, NextResponse } from 'next/server';
+import { NextResponse } from 'next/server';
 import { getOAuthConfig } from '../../../lib/oauth/config';
 
-export async function GET(request: NextRequest) {
+export async function GET() {
   try {
     const config = getOAuthConfig();
     
-    // MCP Protected Resource Metadata according to RFC 8707
-    const protectedResourceMetadata = {
-      // Resource server identifier (canonical URI of this MCP server)
+    const metadata = {
+      // Resource server identifier
       resource: config.issuer,
       
-      // Authorization server that issues tokens for this resource
+      // Authorization server that protects this resource
       authorization_servers: [config.issuer],
       
-      // Scopes supported by this protected resource
+      // MCP server URL (Claude.ai connects to base URL)
+      server_url: config.issuer,
+      mcp_server: config.issuer,
+      
+      // Transport endpoints
+      transport_endpoints: [
+        {
+          type: "http", 
+          url: `${config.issuer}/api/mcp`,
+          methods: ["POST"]
+        },
+        {
+          type: "sse",
+          url: `${config.issuer}/api/sse`, 
+          methods: ["GET", "POST"]
+        }
+      ],
+      
+      // Supported scopes for this resource server
       scopes_supported: config.supportedScopes,
       
-      // Token introspection endpoint (optional)
-      introspection_endpoint: `${config.issuer}/api/oauth/introspect`,
+      // Bearer token usage
+      bearer_methods_supported: ['header'],
+      bearer_locations_supported: ['authorization'],
       
-      // Supported authentication methods for accessing this resource
-      resource_documentation: `${config.issuer}/docs/mcp-authorization`,
+      // Resource server capabilities
+      capabilities: {
+        tools: 18,
+        resources: true,
+        prompts: true,
+        notifications: false,
+        progress: false,
+        logging: true
+      },
       
-      // MCP-specific metadata
-      mcp: {
-        version: "2025-06-18",
-        transport: "http",
-        features: {
-          tools: true,
-          resources: true,
-          prompts: true
+      // Service information
+      service_documentation: `${config.issuer}/docs`,
+      service_name: 'Industrial MCP Server',
+      service_version: '2.0.0',
+      
+      // Database information
+      databases: [
+        {
+          type: 'Neo4j',
+          name: 'Knowledge Graph',
+          description: 'Industrial/operational data and organizational structures'
+        },
+        {
+          type: 'MySQL',
+          name: 'Analytics Database', 
+          description: 'Matomo web analytics and visitor intelligence'
         }
-      }
+      ]
     };
     
     console.log('📋 Serving MCP Protected Resource Metadata');
     
-    return NextResponse.json(protectedResourceMetadata, {
+    return NextResponse.json(metadata, {
       status: 200,
       headers: {
         'Content-Type': 'application/json',
         'Cache-Control': 'public, max-age=3600', // Cache for 1 hour
-        'Access-Control-Allow-Origin': '*',
-        'Access-Control-Allow-Headers': 'Authorization, Content-Type'
       }
     });
-    
   } catch (error) {
     console.error('❌ Error serving protected resource metadata:', error);
     
     return NextResponse.json({
       error: 'server_error',
-      error_description: 'Unable to serve protected resource metadata'
+      error_description: 'Failed to retrieve protected resource metadata'
     }, { 
       status: 500,
       headers: {
@@ -65,17 +95,4 @@ export async function GET(request: NextRequest) {
       }
     });
   }
-}
-
-// Handle CORS preflight
-export async function OPTIONS(request: NextRequest) {
-  return new NextResponse(null, {
-    status: 200,
-    headers: {
-      'Access-Control-Allow-Origin': '*',
-      'Access-Control-Allow-Methods': 'GET, OPTIONS',
-      'Access-Control-Allow-Headers': 'Authorization, Content-Type',
-      'Access-Control-Max-Age': '86400'
-    }
-  });
 }
